# specs/001-crud-empleado/contracts/empleados-openapi.yaml
openapi: 3.0.3
info:
  title: API CRUD Empleados
  version: 1.0.0
  description: Contrato API para alta, consulta, listado, actualización y eliminación de empleados con clave generada `EMP-<autonumérico>`.
servers:
  - url: http://localhost:8080
security:
  - basicAuth: []
paths:
  /api/v1/empleados:
    post:
      summary: Crear empleado
      operationId: crearEmpleado
      tags: [Empleados]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmpleadoCreateRequest'
      responses:
        '201':
          description: Empleado creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoResponse'
        '400':
          description: Error de validación o formato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autenticado
    get:
      summary: Listar empleados
      operationId: listarEmpleados
      tags: [Empleados]
      parameters:
        - name: page
          in: query
          required: false
          description: Número de página (base 0)
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          description: Tamaño de página (por defecto 5)
          schema:
            type: integer
            minimum: 1
            default: 5
      responses:
        '200':
          description: Lista paginada de empleados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoPageResponse'
        '401':
          description: No autenticado
  /api/v1/empleados/{clave}:
    parameters:
      - name: clave
        in: path
        required: true
        description: Clave del empleado con formato EMP-<autonumérico>
        schema:
          type: string
          pattern: '^EMP-[0-9]+$'
    get:
      summary: Consultar empleado por clave
      operationId: obtenerEmpleadoPorClave
      tags: [Empleados]
      responses:
        '200':
          description: Empleado encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoResponse'
        '400':
          description: Formato de clave inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Empleado no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autenticado
    put:
      summary: Actualizar empleado
      operationId: actualizarEmpleado
      tags: [Empleados]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmpleadoUpdateRequest'
      responses:
        '200':
          description: Empleado actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoResponse'
        '400':
          description: Error de validación o formato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Empleado no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autenticado
    delete:
      summary: Eliminar empleado
      operationId: eliminarEmpleado
      tags: [Empleados]
      responses:
        '204':
          description: Empleado eliminado
        '400':
          description: Formato de clave inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Empleado no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autenticado
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    EmpleadoCreateRequest:
      type: object
      required: [nombre, direccion, telefono]
      properties:
        nombre:
          type: string
          maxLength: 100
          example: Ana Pérez
        direccion:
          type: string
          maxLength: 100
          example: Calle Centro 123
        telefono:
          type: string
          maxLength: 100
          example: 5551234567
    EmpleadoUpdateRequest:
      type: object
      required: [nombre, direccion, telefono]
      properties:
        nombre:
          type: string
          maxLength: 100
          example: Ana Pérez López
        direccion:
          type: string
          maxLength: 100
          example: Av. Reforma 456
        telefono:
          type: string
          maxLength: 100
          example: 5557654321
    EmpleadoResponse:
      type: object
      required: [clave, nombre, direccion, telefono]
      properties:
        clave:
          type: string
          pattern: '^EMP-[0-9]+$'
          example: EMP-1001
        nombre:
          type: string
          maxLength: 100
        direccion:
          type: string
          maxLength: 100
        telefono:
          type: string
          maxLength: 100
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [VALIDACION, FORMATO_CLAVE_INVALIDO, CLAVE_NO_EDITABLE, NO_ENCONTRADO]
        message:
          type: string
        details:
          type: array
          items:
            type: string
    EmpleadoPageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/EmpleadoResponse'
        page:
          type: integer
          minimum: 0
          example: 0
        size:
          type: integer
          minimum: 1
          example: 5
        totalElements:
          type: integer
          minimum: 0
          example: 1
        totalPages:
          type: integer
          minimum: 0
          example: 1